fixing commands with warnings

+++ HW1/BuiltInCommands.cpp
@@ -229,7 +229,7 @@ void AliasCommand::execute(){
         std::stringstream arg_buffer;
         std::string line = line_s.str();
         
-        for (int i = 0; i < line.size(); i++){
+        for (size_t i = 0; i < line.size(); i++){
             char ch = line[i];
 
             if (!long_arg){
